sd2-stocksradar/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alpha-vantage==2.3.1
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.2
attrs==22.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1670575068857/work
contourpy==1.0.7
cryptography==39.0.2
cssselect==1.2.0
cycler==0.11.0
dateparser==1.1.7
debugpy @ file:///D:/bld/debugpy_1674522557927/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
et-xmlfile==1.1.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fake-useragent==1.1.1
fastjsonschema==2.16.3
feedparser==6.0.10
fonttools==4.39.2
frozendict==2.3.5
frozenlist==1.3.3
html5lib==1.1
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1672612343532/work
ipykernel @ file:///D:/bld/ipykernel_1678143566265/work
ipython @ file:///D:/bld/ipython_1677617262387/work
ipywidgets==8.0.4
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///D:/bld/jupyter_core_1675110219617/work
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
multidict==6.0.4
multitasking==0.0.11
nbformat==5.7.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.2
openpyxl==3.1.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
pandas-datareader @ git+https://github.com/raphi6/pandas-datareader.git@ea66d6b981554f9d0262038aef2106dda7138316
pandas-ta==0.3.14b0
parse==1.19.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1678652433955/work
plotly==5.13.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil @ file:///D:/bld/psutil_1667886071590/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser==2.21
pycryptodome @ file:///D:/bld/pycryptodome_1669581791697/work
pycryptodomex @ file:///D:/bld/pycryptodomex_1669581405476/work
pyee==8.2.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyparsing==3.0.9
pyppeteer==1.0.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
PyQt6-WebEngine-Qt6==6.4.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyquery==2.0.0
pyrsistent==0.19.3
python-binance==1.0.17
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyzmq @ file:///D:/bld/pyzmq_1678777053265/work
regex==2022.10.31
requests==2.28.2
requests-html==0.10.0
sgmllib3k==1.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
ta==0.10.2
tenacity==8.2.2
tornado @ file:///D:/bld/tornado_1666788735597/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.15
w3lib==2.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
websockets==10.4
widgetsnbextension==4.0.5
yahoo-fin==0.8.9.1
yarl==1.8.2
yfinance==0.2.12
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
